requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
pyarrow==14.0.1